refactor: move context from store to providers in components folder
