+++ apps/web/src/components/auth/CreateOrganizationForm.tsx
@@ -7,7 +7,7 @@ import { IJwtPayload } from '@novu/shared';
 
 import { Button, Input } from '../../design-system';
 import { api } from '../../api/api.client';
-import { useAuthContext } from '../../store/authContext';
+import { useAuthContext } from '../providers/AuthProvider';
 import { useVercelIntegration } from '../../hooks/useVercelIntegration';
 import { useVercelParams } from '../../hooks/useVercelParams';
 import { ROUTES } from '../../constants/routes.enum';
