+++ apps/web/src/components/auth/useAcceptInvite.ts
@@ -5,7 +5,7 @@ import { showNotification } from '@mantine/notifications';
 import * as Sentry from '@sentry/react';
 
 import { api } from '../../api/api.client';
-import { useAuthContext } from '../../store/authContext';
+import { useAuthContext } from '../providers/AuthProvider';
 import { applyToken } from '../../hooks/useAuthController';
 
 export function useAcceptInvite() {
