+++ apps/web/src/components/layout/AppLayout.tsx
@@ -10,7 +10,7 @@ import { INTERCOM_APP_ID } from '../../config';
 import { HEADER_HEIGHT } from './constants';
 import { RequiredAuth } from './RequiredAuth';
 import { SpotLight } from '../utils/Spotlight';
-import { SpotLightProvider } from '../../store/spotlightContext';
+import { SpotLightProvider } from '../providers/SpotlightProvider';
 
 export function AppLayout() {
   return (
