+++ apps/web/src/components/providers/SpotlightProvider.tsx
