+++ apps/web/src/components/utils/Spotlight.tsx
@@ -2,7 +2,7 @@ import { SpotlightProvider } from '@mantine/spotlight';
 import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { Activity, Bolt, Box, Settings, Repeat, Team, Brand, Chat } from '../../design-system/icons';
-import { useSpotlightContext } from '../../store/spotlightContext';
+import { useSpotlightContext } from '../providers/SpotlightProvider';
 import { ROUTES } from '../../constants/routes.enum';
 
 export const SpotLight = ({ children }) => {
