+++ apps/web/src/hooks/useEnvController.ts
@@ -4,7 +4,7 @@ import { IEnvironment } from '@novu/shared';
 
 import { getCurrentEnvironment, getMyEnvironments } from '../api/environment';
 import { api } from '../api/api.client';
-import { useAuthContext } from '../store/authContext';
+import { useAuthContext } from '../components/providers/AuthProvider';
 import { QueryKeys } from '../api/query.keys';
 
 export type EnvironmentContext = {
