+++ apps/web/src/hooks/useNovu.ts
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 import { APP_ID, ENV, WIDGET_EMBED_PATH } from '../config';
-import { useAuthContext } from '../store/authContext';
+import { useAuthContext } from '../components/providers/AuthProvider';
 
 export function useNovu() {
   const authContext = useAuthContext();
