+++ apps/web/src/pages/auth/InvitationPage.tsx
@@ -9,7 +9,7 @@ import AuthLayout from '../../components/layout/components/AuthLayout';
 import AuthContainer from '../../components/layout/components/AuthContainer';
 import { SignUpForm } from '../../components/auth/SignUpForm';
 import { colors, Text, Button } from '../../design-system';
-import { useAuthContext } from '../../store/authContext';
+import { useAuthContext } from '../../components/providers/AuthProvider';
 import { useAcceptInvite } from '../../components/auth/useAcceptInvite';
 import { LoginForm } from '../../components/auth/LoginForm';
