+++ apps/web/src/pages/auth/LoginPage.tsx
@@ -3,14 +3,14 @@ import { useNavigate, useSearchParams } from 'react-router-dom';
 import jwtDecode from 'jwt-decode';
 import { IJwtPayload } from '@novu/shared';
 
-import { useAuthContext } from '../../store/authContext';
+import { useAuthContext } from '../../components/providers/AuthProvider';
 import { LoginForm } from '../../components/auth/LoginForm';
 import AuthLayout from '../../components/layout/components/AuthLayout';
 import AuthContainer from '../../components/layout/components/AuthContainer';
 import { useVercelIntegration } from '../../hooks/useVercelIntegration';
 import SetupLoader from '../../components/auth/SetupLoader';
 import { useVercelParams } from '../../hooks/useVercelParams';
-import { useSegment } from '../../hooks/useSegment';
+import { useSegment } from '../../components/providers/SegmentProvider';
 import { useAcceptInvite } from '../../components/auth/useAcceptInvite';
 import { useBlueprint } from '../../hooks/useBlueprint';
 import { ROUTES } from '../../constants/routes.enum';
