+++ apps/web/src/pages/invites/MembersInvitePage.tsx
@@ -20,7 +20,7 @@ import {
 import { MembersTable } from '../../components/invites/MembersTable';
 import { Button, Input } from '../../design-system';
 import { Invite } from '../../design-system/icons';
-import { useAuthContext } from '../../store/authContext';
+import { useAuthContext } from '../../components/providers/AuthProvider';
 import { parseUrl } from '../../utils/routeUtils';
 import { ROUTES } from '../../constants/routes.enum';
